server/MuseTalk/requirements.txt
absl-py==2.1.0
accelerate==0.28.0
addict==2.4.0
aiofiles==23.2.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
astunparse==1.6.3
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chumpy==0.70
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.1
cycler==0.12.1
Cython==3.0.12
decorator==5.1.1
diffusers==0.32.2
distro==1.9.0
fastapi==0.115.8
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.14.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.67.0
gradio==5.16.0
gradio_client==1.7.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
Jinja2==3.1.5
jmespath==0.10.0
json-tricks==3.17.3
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
lumaai==1.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.3.2
mmcv==2.1.0
mmdet==3.2.0
mmengine==0.10.6
mmpose==1.3.2
model-index==0.1.11
moviepy==2.1.2
mpmath==1.3.0
munkres==1.1.4
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
opt_einsum==3.4.0
optree==0.14.0
ordered-set==4.1.0
orjson==3.10.15
oss2==2.17.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
progress==1.6
proto-plus==1.26.0
protobuf==4.25.6
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2023.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
rsa==4.9
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.5.2
scipy==1.15.1
semantic-version==2.10.0
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
spaces==0.32.0
starlette==0.45.3
static_ffmpeg==2.8
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.16.0
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.5.0
terminaltables==3.1.10
tf_keras-nightly==2.16.0.dev2024021410
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.65.2
transformers==4.48.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xtcocotools==1.14.3
yapf==0.43.0
zipp==3.21.0
